Refactor: move quadrant logic out of table.tsx
